セットの周期の内容を別ファイルに切り出し
