+++ src/App.js
@@ -2,18 +2,7 @@ import React, { useState, useEffect } from "react";
 import "./App.css";
 import Button from "./Button";
 import Timer from "./Timer";
-
-const phases = [
-  { label: "Work 1 (25分)", duration: 1500 },
-  { label: "Break 1 (5分)", duration: 300 },
-  { label: "Work 2 (25分)", duration: 1500 },
-  { label: "Break 2 (5分)", duration: 300 },
-  { label: "Work 3 (25分)", duration: 1500 },
-  { label: "Break 3 (5分)", duration: 300 },
-  { label: "Work 4 (25分)", duration: 1500 },
-  { label: "Break 4 (5分)", duration: 300 },
-  { label: "Long Break (15分)", duration: 900 },
-];
+import phases from "./phases";
 
 function App() {
   const [currentPhaseIndex, setCurrentPhaseIndex] = useState(0);
