+++ src/phases.js
@@ -1,0 +1,13 @@
+const phases = [
+  { label: "Work 1 (25分)", duration: 1500 },
+  { label: "Break 1 (5分)", duration: 300 },
+  { label: "Work 2 (25分)", duration: 1500 },
+  { label: "Break 2 (5分)", duration: 300 },
+  { label: "Work 3 (25分)", duration: 1500 },
+  { label: "Break 3 (5分)", duration: 300 },
+  { label: "Work 4 (25分)", duration: 1500 },
+  { label: "Break 4 (5分)", duration: 300 },
+  { label: "Long Break (15分)", duration: 900 },
+];
+
+export default phases;
